fix: unify MMR rank bands and resolve all clippy warnings
Consolidate MMR classification to single source of truth in router.rs
(5 tiers: Rookie/Bronze/Silver/Gold/Elite). auth.rs now delegates to
router::classify_mmr_band() instead of maintaining a separate 4-tier
system. Resolve all 113 clippy warnings across server and stress-client
crates via --fix and manual fixes (Iterator::find, .clamp(), .or_default(),
collapsible if-let, redundant rebindings, missing is_empty, etc.).

+++ server/src/server/instance/bot_management.rs
@@ -345,7 +345,7 @@ impl MassiveGameServer {
 
             let potential_spawns_for_team: Vec<Vec2> = team_spawn_areas
                 .iter()
-                .filter(|(_, sp_team_id)| *sp_team_id == team_id as u8)
+                .filter(|(_, sp_team_id)| *sp_team_id == team_id)
                 .map(|(pos, _)| *pos)
                 .collect();
 
@@ -366,7 +366,7 @@ impl MassiveGameServer {
                 self.respawn_manager.get_respawn_position(
                     self,
                     &Arc::new(bot_player_id_str.clone()),
-                    Some(team_id as u8),
+                    Some(team_id),
                     &[],
                 )
             };
